This PR contains a few separate optimizations:
* Added mixed_stack_heap_allocator as a way to optimize GradientSurfaceDataComponent so that it performs better on single-point queries.
* Fixed the terrain benchmarks so that building a list of points to query uses a properly-sized container.
* Moved the mutex out of FindBestAreaEntityAtPosition to improve timings on bulk queries.

GradientSurfaceData benchmarks - single query improved without hurting the timings of the bulk queries:

before:
```
---------------------------------------------------------------------------------------------------------------------
Benchmark                                                                           Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------------
GradientSurfaceData/BM_GetSurfacePoints/1024                                      568 ms          578 ms            1
GradientSurfaceData/BM_GetSurfacePoints/2048                                     2234 ms         2234 ms            1
GradientSurfaceData/BM_GetSurfacePointsFromRegion/1024                           66.5 ms         66.8 ms           11
GradientSurfaceData/BM_GetSurfacePointsFromRegion/2048                            265 ms          266 ms            2
GradientSurfaceData/BM_GetSurfacePointsFromList/1024                             66.7 ms         66.8 ms           11
GradientSurfaceData/BM_GetSurfacePointsFromList/2048                              273 ms          276 ms            3
```

after:
```
-------------------------------------------------------------------------------------------------
Benchmark                                                       Time             CPU   Iterations
-------------------------------------------------------------------------------------------------
GradientSurfaceData/BM_GetSurfacePoints/1024                  418 ms          414 ms            2
GradientSurfaceData/BM_GetSurfacePoints/2048                 1639 ms         1641 ms            1
GradientSurfaceData/BM_GetSurfacePointsFromRegion/1024       67.7 ms         67.2 ms           10
GradientSurfaceData/BM_GetSurfacePointsFromRegion/2048        270 ms          273 ms            2
GradientSurfaceData/BM_GetSurfacePointsFromList/1024         66.4 ms         66.8 ms           11
GradientSurfaceData/BM_GetSurfacePointsFromList/2048          267 ms          271 ms            3
```

Terrain benchmarks - bulk query improvements :

before:
```
-----------------------------------------------------------------------------------------------------------------------
Benchmark                                                                             Time             CPU   Iterations
-----------------------------------------------------------------------------------------------------------------------
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegion/1024/1/0                      141 ms          145 ms            4
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegion/2048/1/0                      638 ms          641 ms            1
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegion/4096/1/0                     2566 ms         2563 ms            1
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegionAsync/1024/1/0                 135 ms        0.000 ms          100
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegionAsync/2048/1/0                 550 ms        0.000 ms           10
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegionAsync/4096/1/0                2567 ms        0.000 ms            1
TerrainSystemBenchmarkFixture/BM_ProcessSurfacePointsList/1024/1/0                  880 ms          875 ms            1
TerrainSystemBenchmarkFixture/BM_ProcessSurfacePointsList/2048/1/0                 3433 ms         3438 ms            1
TerrainSystemBenchmarkFixture/BM_ProcessSurfacePointsListAsync/1024/1/0             759 ms         3.13 ms           10
TerrainSystemBenchmarkFixture/BM_ProcessSurfacePointsListAsync/2048/1/0            3430 ms         93.8 ms            1
```

after:
```
-----------------------------------------------------------------------------------------------------------------------
Benchmark                                                                             Time             CPU   Iterations
-----------------------------------------------------------------------------------------------------------------------
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegion/1024/1/0                     97.1 ms         99.0 ms            6
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegion/2048/1/0                      430 ms          430 ms            2
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegion/4096/1/0                     1889 ms         1891 ms            1
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegionAsync/1024/1/0                92.9 ms        0.000 ms          100
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegionAsync/2048/1/0                 382 ms        0.000 ms           10
TerrainSystemBenchmarkFixture/BM_ProcessHeightsRegionAsync/4096/1/0                1536 ms        0.000 ms           10
TerrainSystemBenchmarkFixture/BM_ProcessSurfacePointsList/1024/1/0                  666 ms          656 ms            1
TerrainSystemBenchmarkFixture/BM_ProcessSurfacePointsList/2048/1/0                 2532 ms         2531 ms            1
TerrainSystemBenchmarkFixture/BM_ProcessSurfacePointsListAsync/1024/1/0             538 ms        0.000 ms           10
TerrainSystemBenchmarkFixture/BM_ProcessSurfacePointsListAsync/2048/1/0            2531 ms         31.3 ms            1
```